requirements.txt
requests
typer
rich
python-dotenv
pyinstaller
